+++ src/index.spec.js
@@ -377,11 +377,10 @@ describe('Module: scancss', () => {
 					anonymousReplacedElements: 0,
 					performanceHacks: {},
 					autoKeyword: 1,
-					explicitDefaultingKeywords: {
-						total: 0,
-						unique: 0,
-						usage: {},
-					},
+					inheritKeyword: 0,
+					initialKeyword: 0,
+					revertKeyword: 0,
+					unsetKeyword: 0,
 					engineTriggers: {
 						composite: {
 							blink: 16,